+++ IlmBase/ImathTest/testLineAlgo.cpp
@@ -344,8 +344,8 @@ testIntersect ()
 
 	    do
 	    {
-		b.x = rand.nextf (0.001, 0.999);
-		b.y = rand.nextf (0.001, 0.999);
+		b.x = static_cast<float>(rand.nextf (0.001, 0.999));
+		b.y = static_cast<float>(rand.nextf (0.001, 0.999));
 		b.z = 1 - b.x - b.y;
 	    }
 	    while (b.x + b.y > 0.999);
@@ -390,8 +390,8 @@ testIntersect ()
 
 	    do
 	    {
-		b.x = rand.nextf (-3, 3);
-		b.y = rand.nextf (-3, 3);
+		b.x = static_cast<float>(rand.nextf (-3, 3));
+		b.y = static_cast<float>(rand.nextf (-3, 3));
 		b.z = 1 - b.x - b.y;
 	    }
 	    while (b.x > -0.001 && b.y > -0.001 && b.z > -0.001);
